General cleanup and making more robust.
Now USB works normally. But when USB is idle for > 2 seconds, it checks if an
SD card is available. And if so, it changes serial speed and starts streaming
data. When USB becomes active again, the SD card stuff is stopped, and serial
speed reverts back to the USB-configured one.

Still a couple bugs:

 - There seems to be a few bytes offset in the sent frames, but otherwise it
   runs stable.

 - The stuff only works when the PC is reading from the USB, for some reason.